# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Institut Lwa-Nzururu - Système de gestion scolaire
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/setup-account:
    post:
      operationId: setupAccount
      tags: [auth]
      summary: Setup account on first login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupAccountBody"
      responses:
        "200":
          description: Account setup complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List all classes
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"

    post:
      operationId: createClass
      tags: [classes]
      summary: Create a class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassBody"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /classes/{id}:
    get:
      operationId: getClass
      tags: [classes]
      summary: Get class by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateClass
      tags: [classes]
      summary: Update class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClassBody"
      responses:
        "200":
          description: Class updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"

    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /archives:
    get:
      operationId: listArchives
      tags: [archives]
      summary: List archived academic years
      responses:
        "200":
          description: List of archives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Archive"

    post:
      operationId: createArchive
      tags: [archives]
      summary: Archive an academic year
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArchiveBody"
      responses:
        "201":
          description: Archive created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Archive"
  /archives/{id}:
    get:
      operationId: getArchive
      tags: [archives]
      summary: Get archive by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Archive found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Archive"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        requiresSetup:
          type: boolean
      required:
        - user
        - requiresSetup
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword
    SetupAccountBody:
      type: object
      properties:
        tempUsername:
          type: string
        tempPassword:
          type: string
        newUsername:
          type: string
        newPassword:
          type: string
        fullName:
          type: string
      required:
        - tempUsername
        - tempPassword
        - newUsername
        - newPassword
        - fullName
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [proviseur, enseignant, titulaire, secretaire]
        classId:
          type: ["integer", "null"]
        className:
          type: ["string", "null"]
        isFirstLogin:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - fullName
        - role
        - isFirstLogin
        - createdAt
...
    Class:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: string
        section:
          type: ["string", "null"]
        room:
          type: ["string", "null"]
        titulaire:
          type: ["string", "null"]
        studentCount:
          type: integer
        academicYear:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - level
        - studentCount
        - academicYear
        - createdAt
    CreateClassBody:
      type: object
      properties:
        name:
          type: string
        level:
          type: string
        section:
          type: ["string", "null"]
        room:
          type: ["string", "null"]
        academicYear:
          type: string
      required:
        - name
        - level
        - academicYear
    UpdateClassBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        level:
          type: ["string", "null"]
        section:
          type: ["string", "null"]
        room:
          type: ["string", "null"]
        academicYear:
          type: ["string", "null"]
...
    Archive:
      type: object
      properties:
        id:
          type: integer
        academicYear:
          type: string
        archivedAt:
          type: string
        totalStudents:
          type: integer
        totalClasses:
          type: integer
        overallPassRate:
          type: number
        notes:
          type: ["string", "null"]
        createdBy:
          type: string
      required:
        - id
        - academicYear
        - archivedAt
        - totalStudents
        - totalClasses
        - overallPassRate
        - createdBy
    CreateArchiveBody:
      type: object
      properties:
        academicYear:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - academicYear
...