# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Institut Lwa-Nzururu - Système de gestion scolaire
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  username: string;
  fullName: string;
  role: UserRole;
  /** @nullable */
  classId?: number | null;
  /** @nullable */
  className?: string | null;
  isFirstLogin: boolean;
  createdAt: string;
}